% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close.all.spectrometers.R
\name{close_all_spectrometers}
\alias{close_all_spectrometers}
\alias{srs_close}
\title{close all open connections to spectrometers}
\usage{
close_all_spectrometers(jwrapper)

srs_close(jwrapper)
}
\arguments{
\item{jwrapper}{a open Wrapper object from Omnidriver}
}
\value{
a numeric value
}
\description{
close all open connections to spectrometers
}
