\name{get.bench}
\alias{get.bench}
\title{Get optical properties of optical bench}
\usage{
  get.bench(jwrapper, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Get optical properties of optical bench
}

