\name{get.boxcar.width}
\alias{get.boxcar.width}
\title{Get current setting of boxcar width (the number of pixels averaged)}
\usage{
  get.boxcar.width(jwrapper, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Get current setting of boxcar width (the number of pixels
  averaged)
}

