\name{get.correct.for.detector.nonlineary}
\alias{get.correct.for.detector.nonlineary}
\title{Get current setting of linearity correction (enabled or not)}
\usage{
  get.correct.for.detector.nonlineary(jwrapper,
    sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Get current setting of linearity correction (enabled or
  not)
}

