\name{get.correct.for.electrical.dark}
\alias{get.correct.for.electrical.dark}
\title{Get current setting for electrical dark correction}
\usage{
  get.correct.for.electrical.dark(jwrapper, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Get current setting for electrical dark correction
}

