\name{get.correct.for.stray.light}
\alias{get.correct.for.stray.light}
\title{Get current setting of stray light correction}
\usage{
  get.correct.for.stray.light(jwrapper, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Get current setting of stray light correction
}

