\name{get.detector}
\alias{get.detector}
\title{Get description of detector}
\usage{
  get.detector(jwrapper, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Get description of detector
}

