\name{get.feature.controller.board.temperature}
\alias{get.feature.controller.board.temperature}
\title{Get feature "controller board temperature"
Before calling this function you first need to check that the feature is supported
by the spectrometer in use.}
\usage{
  get.feature.controller.board.temperature(jwrapper,
    sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Get feature "controller board temperature" Before calling
  this function you first need to check that the feature is
  supported by the spectrometer in use.
}

