\name{get.feature.detector.board.temperature}
\alias{get.feature.detector.board.temperature}
\title{Get feature "detector temperature"
Before calling this function you first need to check that the feature is supported
by the spectrometer in use.}
\usage{
  get.feature.detector.board.temperature(jwrapper,
    sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Get feature "detector temperature" Before calling this
  function you first need to check that the feature is
  supported by the spectrometer in use.
}

