\name{get.feature.pixel.binning}
\alias{get.feature.pixel.binning}
\title{Get feature "pixel binning"
Before calling this function you first need to check that the feature is supported
by the spectrometer in use.}
\usage{
  get.feature.pixel.binning(jwrapper, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Get feature "pixel binning" Before calling this function
  you first need to check that the feature is supported by
  the spectrometer in use.
}

