\name{get.spectrum}
\alias{get.spectrum}
\title{Get a spectrum from the spectrometer}
\usage{
  get.spectrum(jwrapper, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Get a spectrum from the spectrometer
}

