% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.api.version.R
\name{get_api_version}
\alias{get_api_version}
\title{Get OmniDriver API version}
\usage{
get_api_version(jwrapper)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}
}
\value{
A character string value of form "1.01.01" or similar.
}
\description{
A function to query the version of OmniDriver.
}
