% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_bench}
\alias{get_bench}
\title{Get optical properties of optical bench}
\usage{
get_bench(jwrapper, sr.index=0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
a value
}
\description{
A function to query the characteristics of the optical bench
from a spectrometer.
}
\keyword{manip}
\keyword{misc}

