% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.correct.for.detector.nonlinearity.R
\name{get_correct_for_detector_nonlineary}
\alias{get_correct_for_detector_nonlineary}
\title{Get current setting of linearity correction (enabled or not)}
\usage{
get_correct_for_detector_nonlineary(jwrapper, sr.index = 0L,
  ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a numeric value
}
\description{
Get current setting of linearity correction (enabled or not)
}
