% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-saturation-threshold.R
\name{get_feature_controller_saturation_threshold}
\alias{get_feature_controller_saturation_threshold}
\title{Get feature "saturation threshold"}
\usage{
get_feature_controller_saturation_threshold(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
an object which provides the desired interface, or null if this 
  feature is not available for this spectrometer.
}
\description{
A function to retrieve an interface that allows to control the saturation
threshold.
}
\note{
Before calling this function you first need to check that the feature 
  is supported by the spectrometer in use by calling 
  \code{\link{is_feature_supported_saturation_threshold}}.
}
