% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-SPI-bus.R
\name{get_feature_SPI_bus}
\alias{get_feature_SPI_bus}
\title{Get feature controller "SPI Bus"}
\usage{
get_feature_SPI_bus(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
an object which provides the desired interface, or null if this
  feature is not available for this spectrometer.
}
\description{
Get a wrapper for accessing this optional feature.
}
\note{
Before calling this function you first need to check that the feature
  is supported by the spectrometer in use by calling
  \code{\link{is_feature_supported_SPI_bus}}.
}
\seealso{
Other Spectrometer I2C- and SPI-bus functions.: 
\code{\link{get_I2C_bus}()},
\code{\link{get_SPI_bytes}()},
\code{\link{get_feature_I2C_bus}()},
\code{\link{is_feature_supported_I2C_bus}()},
\code{\link{is_feature_supported_SPI_bus}()}
}
\concept{Spectrometer I2C- and SPI-bus functions.}
