% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.firmware.model.R
\name{get_firmware_model}
\alias{get_firmware_model}
\title{Get name of spectrometer}
\usage{
get_firmware_model(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
A character string value of form "USB2000" or similar.
}
\description{
A function to query the version of the firmware from a spectrometer.
}
\note{
Only supported in API version >= 2.40, in earlier API versions an error
  is triggered if use is attempted.
}
