% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.firmware.version.R
\name{get_firmware_version}
\alias{get_firmware_version}
\title{Get spectrometer firmware version}
\usage{
get_firmware_version(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
A character string value of form "1.01.01" or similar.
}
\description{
A function to query the version of the firmware from a spectrometer.
}
