% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.integration.time.R
\name{get_integration_time}
\alias{get_integration_time}
\title{Get integration time}
\usage{
get_integration_time(jwrapper, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
current integration time setting, in units of microseconds.
}
\description{
A function to query the integration time in use
from a spectrometer.
}
