% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.last.exception.R
\name{get_last_exception}
\alias{get_last_exception}
\title{Get OmniDriver last exception}
\usage{
get_last_exception(jwrapper)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}
}
\value{
A character string String with a description of the most recent error
  exception, or "no exception" if no error has been encountered..
}
\description{
A function to retrieve the most recent Java exception triggered by OmniDrive 
as a character string.
}
