% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.maximum.integration.time.R
\name{get_maximum_integration_time}
\alias{get_maximum_integration_time}
\title{Get maximum integration time}
\usage{
get_maximum_integration_time(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
maximum integration time setting, in units of microseconds.
}
\description{
A function to query the maximum integration time that can be used
from a spectrometer.
}
