% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.maximum.intensity.R
\name{get_maximum_intensity}
\alias{get_maximum_intensity}
\title{Get maximum intensity}
\usage{
get_maximum_intensity(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
maximum value per pixel, in units of counts???.
}
\description{
A function to query the maximum integration time that can be used
from a spectrometer.
}
