% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.minimum.integration.time.R
\name{get_minimum_integration_time}
\alias{get_minimum_integration_time}
\title{Get minimum integration time}
\usage{
get_minimum_integration_time(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
minimum integration time setting, in units of microseconds.
}
\description{
A function to query the minimum integration time that can be used
from a spectrometer.
}
