% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.name.R
\name{get_name}
\alias{get_name}
\title{Get name of spectrometer}
\usage{
get_name(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
a value
}
\description{
A function to query the name (e.g. USB2000)
from a spectrometer.
}
