% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.number.of.channels.R
\name{get_number_of_channels}
\alias{get_number_of_channels}
\title{Get number of channels}
\usage{
get_number_of_channels(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
a numeric value
}
\description{
A function to query the number of channels
from a spectrometer.
}
