% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.number.of.dark.pixels.R
\name{get_number_of_dark_pixels}
\alias{get_number_of_dark_pixels}
\title{Get number of dark pixels}
\usage{
get_number_of_dark_pixels(jwrapper, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a numeric value with the total number of pixels (ie. CCD elements)
  provided by this spectrometer, including any dark or bevel (unused) pixels.
}
\description{
Function to query the number of pixels in the CCD array.
}
