% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.scans.to.avg.R
\name{get_scans_to_average}
\alias{get_scans_to_average}
\alias{get_scans_to_avg}
\title{Get "number of scans to average"}
\usage{
get_scans_to_average(jwrapper, sr.index = 0L, ch.index = 0L)

get_scans_to_avg(jwrapper, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a numeric value
}
\description{
Get the setting "number of scans to average" currently in use by the
addressed spectrometer channel.
}
