% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-time-delta-since.R
\name{get_seconds_time_delta_since}
\alias{get_seconds_time_delta_since}
\alias{get_nano_time_delta_since}
\alias{get_micro_time_delta_since}
\alias{get_milli_time_delta_since}
\title{Get time difference since a time stamp}
\usage{
get_seconds_time_delta_since(then)

get_nano_time_delta_since(then)

get_micro_time_delta_since(then)

get_milli_time_delta_since(then)
}
\arguments{
\item{then}{HighResTimeStamp defined in OmniDriver API and retrieved
or constructed with other OmniDriver API methods. Not vectorized!}
}
\value{
a numeric vector of length one giving the time difference in
  seconds, milliseconds, microseconds or nanoseconds.
}
\description{
Compute the time difference between a high speed time stamp and current time.
}
\note{
Nanosecond timing and normal timing are not coincident, but as both
are stored in parallel, the differences should differ only in their
resolution.
}
\seealso{
Other high speed acquisition functions: 
\code{\link{get_seconds_time_delta}()},
\code{\link{highSpdAcq_allocate_buffer}()},
\code{\link{highSpdAcq_get_number_of_spectra_acquired}()},
\code{\link{highSpdAcq_get_spectrum}()},
\code{\link{highSpdAcq_get_time_stamp}()},
\code{\link{highSpdAcq_start_acquisition}()},
\code{\link{oo_timestamp_to_string}()}

Other high speed acquisition functions: 
\code{\link{get_seconds_time_delta}()},
\code{\link{highSpdAcq_allocate_buffer}()},
\code{\link{highSpdAcq_get_number_of_spectra_acquired}()},
\code{\link{highSpdAcq_get_spectrum}()},
\code{\link{highSpdAcq_get_time_stamp}()},
\code{\link{highSpdAcq_start_acquisition}()},
\code{\link{oo_timestamp_to_string}()}

Other high speed acquisition functions: 
\code{\link{get_seconds_time_delta}()},
\code{\link{highSpdAcq_allocate_buffer}()},
\code{\link{highSpdAcq_get_number_of_spectra_acquired}()},
\code{\link{highSpdAcq_get_spectrum}()},
\code{\link{highSpdAcq_get_time_stamp}()},
\code{\link{highSpdAcq_start_acquisition}()},
\code{\link{oo_timestamp_to_string}()}

Other high speed acquisition functions: 
\code{\link{get_seconds_time_delta}()},
\code{\link{highSpdAcq_allocate_buffer}()},
\code{\link{highSpdAcq_get_number_of_spectra_acquired}()},
\code{\link{highSpdAcq_get_spectrum}()},
\code{\link{highSpdAcq_get_time_stamp}()},
\code{\link{highSpdAcq_start_acquisition}()},
\code{\link{oo_timestamp_to_string}()}
}
\concept{high speed acquisition functions}
