% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.serial.number.R
\name{get_serial_number}
\alias{get_serial_number}
\title{Get serial number of spectrometer}
\usage{
get_serial_number(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
a value
}
\description{
A function to query the serial number
from a spectrometer.
}
