% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.spectrum.R
\name{get_spectrum}
\alias{get_spectrum}
\title{Get a spectrum from the spectrometer}
\usage{
get_spectrum(jwrapper, sr.index = 0L, ch.index = NULL)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a numeric value
}
\description{
Get a spectrum from the spectrometer
}
\keyword{manip}
\keyword{misc}

