% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-SPI-bus.R
\name{get_SPI_bytes}
\alias{get_SPI_bytes}
\title{Talk to "SPI bus"}
\usage{
get_SPI_bytes(SPI.bus.wrapper, message, num.bytes)
}
\arguments{
\item{SPI.bus.wrapper}{feature wrapper as returned by function
\code{rOmniDriver::get_feature_controller_SPI_bus}}

\item{message}{bytes, array of bytes with maximum length of 62.}

\item{num.bytes}{numeric Number of bytes, 1..62.}
}
\value{
After writing to SPI, the spectrometer immediately reads from it. The
  resulting Byte array is returned.
}
\description{
Send a packet of bytes over the SPI bus and read the result.
}
\details{
Several spectrometers have an I2C bus accesible through the auxiliary 
connector which can be used to control accesories. The I2C bus in Ocean
Insight spectrometers is not intended to be used to control the spectrometer
or retrieve data but instead to control and access other accessories with
the spectrometer acting as middle man.
}
\seealso{
Other Spectrometer I2C- and SPI-bus functions.: 
\code{\link{get_I2C_bus}()},
\code{\link{get_feature_I2C_bus}()},
\code{\link{get_feature_SPI_bus}()},
\code{\link{is_feature_supported_I2C_bus}()},
\code{\link{is_feature_supported_SPI_bus}()}
}
\concept{Spectrometer I2C- and SPI-bus functions.}
