% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.wavelengths.R
\name{get_wavelengths}
\alias{get_wavelengths}
\title{Get wavelengths from spectrometer}
\usage{
get_wavelengths(jwrapper, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
A numeric (double) vector with the calculated wavelength at each pixel.
}
\description{
Function to get wavelengths corresponding to the pixels in the CCD array.
}
