% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.wrapper.extensions.R
\name{get_wrapper_extensions}
\alias{get_wrapper_extensions}
\title{Get OmniDriver wrapper extensions}
\usage{
get_wrapper_extensions(jwrapper)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}
}
\value{
A character string value of form "1.01.01" or similar.
}
\description{
A function to query enable the WrapperExtensions class.
}
