% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highSpdAcq-AllocateBuffer.R
\name{highSpdAcq_allocate_buffer}
\alias{highSpdAcq_allocate_buffer}
\title{Allocate memory buffer for high speed acquisition}
\usage{
highSpdAcq_allocate_buffer(jwrapper, sr.index = 0L, number.of.spectra = 100L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{number.of.spectra}{integer The number of spectra to be acquired.}
}
\value{
a numeric value
}
\description{
Pre-allocates a buffer. A call to this function should always precede a call
to \code{highSpdAcq_start_acquisition()}
}
\seealso{
Other high speed acquisition functions: 
\code{\link{get_seconds_time_delta}()},
\code{\link{get_seconds_time_delta_since}()},
\code{\link{highSpdAcq_get_number_of_spectra_acquired}()},
\code{\link{highSpdAcq_get_spectrum}()},
\code{\link{highSpdAcq_get_time_stamp}()},
\code{\link{highSpdAcq_start_acquisition}()},
\code{\link{oo_timestamp_to_string}()}
}
\concept{high speed acquisition functions}
