% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highSpdAcq-GetNumberOfSpectraAcquired.R
\name{highSpdAcq_get_number_of_spectra_acquired}
\alias{highSpdAcq_get_number_of_spectra_acquired}
\title{Get the number of spectra acquired at high speed}
\usage{
highSpdAcq_get_number_of_spectra_acquired(jwrapper)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}
}
\value{
a numeric value
}
\description{
The number of spectra acquired can be smaller than the intended one, and
this function retrieved the true number of spectra available in the buffer
after the acquisition has finished. A call to this function can be used
after a call to \code{highSpdAcq_start_acquisition()} returns.
}
\seealso{
Other high speed acquisition functions: 
\code{\link{get_seconds_time_delta_since}()},
\code{\link{get_seconds_time_delta}()},
\code{\link{highSpdAcq_allocate_buffer}()},
\code{\link{highSpdAcq_get_spectrum}()},
\code{\link{highSpdAcq_get_time_stamp}()},
\code{\link{highSpdAcq_start_acquisition}()},
\code{\link{oo_timestamp_to_string}()}
}
\concept{high speed acquisition functions}
