% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highSpdAcq-GetTimeStamp.R
\name{highSpdAcq_get_time_stamp}
\alias{highSpdAcq_get_time_stamp}
\title{Get the time stamp of one spectrum acquired at high speed}
\usage{
highSpdAcq_get_time_stamp(jwrapper, spectrum.number = 1)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{spectrum.number}{integer The index into the sequence of spectra
most recently acquired at high speed and available in the memory buffer.}
}
\value{
A \code{HighResTimeStamp} object.
}
\description{
Retrieve the time stamp for a single, previously measured, spectrum from the
buffer. A call to this function can be used after a call to
\code{highSpdAcq_start_acquisition()} returns.
}
\seealso{
Other high speed acquisition functions: 
\code{\link{get_seconds_time_delta_since}()},
\code{\link{get_seconds_time_delta}()},
\code{\link{highSpdAcq_allocate_buffer}()},
\code{\link{highSpdAcq_get_number_of_spectra_acquired}()},
\code{\link{highSpdAcq_get_spectrum}()},
\code{\link{highSpdAcq_start_acquisition}()},
\code{\link{oo_timestamp_to_string}()}
}
\concept{high speed acquisition functions}
