% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highSpdAcq-StartAcquisition.R
\name{highSpdAcq_start_acquisition}
\alias{highSpdAcq_start_acquisition}
\title{Start high speed acquisition}
\usage{
highSpdAcq_start_acquisition(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
a numeric value
}
\description{
Acquires a series of spectra into a preallocated buffer. A call to this 
function should be always preceded by a call to 
\code{highSpdAcq_allocate_buffer()}.
}
\seealso{
Other high speed acquisition functions: 
\code{\link{get_seconds_time_delta_since}()},
\code{\link{get_seconds_time_delta}()},
\code{\link{highSpdAcq_allocate_buffer}()},
\code{\link{highSpdAcq_get_number_of_spectra_acquired}()},
\code{\link{highSpdAcq_get_spectrum}()},
\code{\link{highSpdAcq_get_time_stamp}()},
\code{\link{oo_timestamp_to_string}()}
}
\concept{high speed acquisition functions}
