% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.api.R
\name{init_api}
\alias{init_api}
\alias{init_srs}
\alias{is_api_enabled}
\alias{init_extended_api}
\alias{is_extended_api_enabled}
\alias{init_highres_time_api}
\alias{is_highres_time_api_enabled}
\title{Initialize the connection to the driver}
\usage{
init_api()

init_srs()

is_api_enabled(jwrapper)

init_extended_api()

is_extended_api_enabled(jwrapperext)

init_highres_time_api()

is_highres_time_api_enabled()
}
\arguments{
\item{jwrapper}{jobjRef A Java Wrapper object from OmniDriver.}

\item{jwrapperext}{jobjRef An "extension" Java Wrapper object from OmniDriver API.}
}
\value{
a wrapper to an instance of a Java object of class "wrapper"

logical

logical
}
\description{
Function \code{init_api()} (and its synonim \code{init_srs()})
  create a Wrapper object for calling Omnidriver functions. This Java object of
  class \code{com.oceanoptics.omnidriver.api.wrapper.Wrapper} is
  later used by all other functions in the rOmniDriver to communicate with 
  the spectrometers through the OmniDriver API from Ocen Insight.

Only one wrapper object can be active at a time. You should call this
function only once and save the returned object as it is used to access the
methods implementing communication with the spectrometer. Clean up by use of
\code{close_srs()} before exiting or before initializing another wrapper to
the same driver.

Function \code{is_api_enableb()} tests if the argument passed to 
\code{jwrapper} has been initialized and if it is connected to an active
instance of the Java wrapper class.

Function \code{init_api_extensions()} enables API extensions
  and returns a wrapper that makes it possible to call additional
  functions.

Function \code{is_extended_api_enabled()} tests if the argument passed to 
\code{jwrapperext} has been initialized and if it is connected to an active
instance of the Java wrapper class.

Function \code{init_api_extensions()} enables API extensions
  and returns a wrapper that makes it possible to call additional
  functions.

Function \code{is_highres_time_api_enabled()} tests if  
wrapper has been initialized and saved to a private environment of the
package.
}
