% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init.srs.R
\name{init_srs}
\alias{init_srs}
\title{Initialize the connection to all currently connected
spectrometers}
\usage{
init_srs()
}
\value{
a wrapper
}
\description{
This creates a Wrapper object using Omnidriver
This object is later used by all other functions to access the spectrometers
}
\details{
You should call this function only once
And cleanup by use of \code{close.srs} before exiting
}
\keyword{manip}
\keyword{misc}

