\name{is.feature.supported.detector.temperature}
\alias{is.feature.supported.detector.temperature}
\title{Checks whether feature "detector temperature" is available
in the spectrometer addressed.}
\usage{
  is.feature.supported.detector.temperature(jwrapper,
    sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Checks whether feature "detector temperature" is
  available in the spectrometer addressed.
}

