\name{is.feature.supported.pixel.binning}
\alias{is.feature.supported.pixel.binning}
\title{Checks whether feature "pixel binning" is available
in the spectrometer addressed.}
\usage{
  is.feature.supported.pixel.binning(jwrapper,
    sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Checks whether feature "pixel binning" is available in
  the spectrometer addressed.
}

