\name{is.saturated}
\alias{is.saturated}
\title{Checks whether the last spectrum acquired is saturated (= clipped).}
\usage{
  is.saturated(jwrapper, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Checks whether the last spectrum acquired is saturated (=
  clipped).
}

