\name{is.spectrum.valid}
\alias{is.spectrum.valid}
\title{Checks whether the last spectrum acquired is valid
(= no errors have occurred during acquisition).}
\usage{
  is.spectrum.valid(jwrapper, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Checks whether the last spectrum acquired is valid (= no
  errors have occurred during acquisition).
}

