\name{is.timeout}
\alias{is.timeout}
\title{Checks whether the last operation has timed out
Usefull when using triggers}
\usage{
  is.timeout(jwrapper, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Checks whether the last operation has timed out Usefull
  when using triggers
}

