% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.number.of.channels.R
\name{is_channel_enabled}
\alias{is_channel_enabled}
\title{Informs whether a channel is enbled or not}
\usage{
is_channel_enabled(jwrapper, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a numeric value
}
\description{
Informs whether a channel of a multichannel spectrometer is enbled or not.
}
