% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-controller-board-temperature.R
\name{is_feature_supported_board_temperature}
\alias{is_feature_supported_board_temperature}
\alias{get_feature_controller_board_temperature}
\alias{get_board_temperature}
\title{Feature "Board temperature"}
\usage{
is_feature_supported_board_temperature(jwrapper, sr.index = 0L)

get_feature_controller_board_temperature(jwrapper, sr.index = 0L)

get_board_temperature(brd.wrapper)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{brd.wrapper}{an open controller wrapper object obtained with
\code{get_feature_controller_board_temperature()}}
}
\value{
a numeric value

an object which provides the desired interface, or null if this
  feature is not available for this spectrometer.
}
\description{
Check whether feature "board temperature" is available in the spectrometer
addressed. Construct a wrapper on the Java controller for the feature. Use a
wrapped controller to query the temperature.
}
\note{
Before calling \code{get_feature_controller_board_temperature} you f
 irst need to check that the feature
  is supported by the spectrometer in use by calling 
  \code{is_feature_supported_board_temperature}.
}
