% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-detector-temperature.R
\name{is_feature_supported_detector_temperature}
\alias{is_feature_supported_detector_temperature}
\alias{get_feature_detector_temperature}
\alias{get_detector_temperature}
\title{Feature "detector temperature"}
\usage{
is_feature_supported_detector_temperature(jwrapper, sr.index = 0L)

get_feature_detector_temperature(jwrapper, sr.index = 0L)

get_detector_temperature(dtr.wrapper)
}
\arguments{
\item{jwrapper}{an open Wrapper object from OmniDriver}

\item{sr.index}{an index to address the spectrometer}

\item{dtr.wrapper}{an open detector wrapper object obtained with
\code{get_feature_detector_temperature()}}
}
\value{
a numeric value

an object which provides the desired interface, or null if this 
  feature is not available for this spectrometer.
}
\description{
Check whether feature "detector temperature" is available in the spectrometer
addressed. Construct a wrapper on the Java controller for the feature. Use a
wrapped controller to query the detector temperature.
}
\note{
Before calling \code{get_feature_detector_temperature} you first need to check that the feature
  is supported by the spectrometer in use#' by calling
  \code{is_feature_supported_detector_temperature}.
}
