% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-internal-trigger.R
\name{is_feature_supported_internal_trigger}
\alias{is_feature_supported_internal_trigger}
\alias{get_feature_controller_internal_trigger}
\alias{get_trigger_period_valid_range}
\alias{get_trigger_period}
\alias{set_trigger_period}
\alias{set_trigger_source}
\alias{get_trigger_source}
\title{Feature "ìnternal trigger"}
\usage{
is_feature_supported_internal_trigger(jwrapper, sr.index = 0L)

get_feature_controller_internal_trigger(jwrapper, sr.index = 0L)

get_trigger_period_valid_range(tgr.wrapper)

get_trigger_period(tgr.wrapper)

set_trigger_period(tgr.wrapper, period)

set_trigger_source(tgr.wrapper, source)

get_trigger_source(tgr.wrapper, char.out = TRUE)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{tgr.wrapper}{feature wrapper as returned by function
\code{rOmniDriver::get_feature_controller_internal_trigger}}

\item{period}{numeric Internal trigger period in seconds.}

\item{source}{character, "on", "off", "external", "internal" or integer
[0..3] or logical.}

\item{char.out}{logical If TRUE return a character string, otherwise return
integer as returned by the API call.}
}
\value{
an integer value

an object which provides the desired interface, or null if this
  feature is not available for this spectrometer.

numeric vector of length three with times in seconds for "min",
  "max" and "step".
}
\description{
Check whether feature "ìnternal trigger" is available in the spectrometer
addressed. Retrieve an interface that allows to control the ìnternal
trigger. Control internal trigger of light sources such as Jaz PX.
}
\note{
Before calling \code{get_feature_controller_internal_trigger} you first need to check that the feature
  is supported by the spectrometer in use by calling 
  \code{is_feature_supported_internal_trigger}.
}
