% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-pixel-binning.R
\name{is_feature_supported_pixel_binning}
\alias{is_feature_supported_pixel_binning}
\title{Is feature "pixel binning" supported?}
\usage{
is_feature_supported_pixel_binning(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
a numeric value
}
\description{
Checks whether feature "pixel binning" is available
in the spectrometer addressed.
}
