% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-saturation-threshold.R
\name{is_feature_supported_saturation_threshold}
\alias{is_feature_supported_saturation_threshold}
\title{Is feature "saturation threshold" supported?}
\usage{
is_feature_supported_saturation_threshold(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
a logical? value
}
\description{
Checks whether feature "saturation threshold" is available in the
spectrometer addressed.
}
