% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-spectrum-type.R
\name{is_feature_supported_spectrum_type}
\alias{is_feature_supported_spectrum_type}
\title{Is feature "spectrum type" supported?}
\usage{
is_feature_supported_spectrum_type(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
a logical value
}
\description{
Checks whether feature "spectrum type" is available
in the spectrometer addressed.
}
