% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-SPI-bus.R
\name{is_feature_supported_SPI_bus}
\alias{is_feature_supported_SPI_bus}
\title{Is feature "SPI Bus" supported?}
\usage{
is_feature_supported_SPI_bus(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
a logical value
}
\description{
Checks whether feature "SPI Bus" is available
in the spectrometer addressed.
}
\seealso{
Other Spectrometer I2C- and SPI-bus functions.: 
\code{\link{get_I2C_bus}()},
\code{\link{get_SPI_bytes}()},
\code{\link{get_feature_I2C_bus}()},
\code{\link{get_feature_SPI_bus}()},
\code{\link{is_feature_supported_I2C_bus}()}
}
\concept{Spectrometer I2C- and SPI-bus functions.}
