% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.saturated.R
\name{is_saturated}
\alias{is_saturated}
\title{Is the most recent spectrum acquired saturated?}
\usage{
is_saturated(jwrapper, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a numeric value
}
\description{
Checks whether the most recently acquired spectrum is saturated (= signal
clipping occured).
}
\note{
Be aware that clipping outside the wavelength range of interest for 
  a given measuremt may not be important and can be ignored. This method
  tests for clipping at any pixel of the array.
}
