% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.spectrum.valid.R
\name{is_spectrum_valid}
\alias{is_spectrum_valid}
\title{Is the most recent spectrum acquired valid?}
\usage{
is_spectrum_valid(jwrapper, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a numeric value
}
\description{
Informs whether the most recently acquired spectrum is valid 
(no communication or similar errors have occurred during acquisition).
}
