% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.USB.timeout.R
\name{is_USB_timeout}
\alias{is_USB_timeout}
\title{Did the last USB operation time out?}
\usage{
is_USB_timeout(jwrapper, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a numeric value
}
\description{
Checks whether the last USB operation has timed out. This is useful when working 
with triggers together with signals used for triggering that may fail or be
delayed.
}
