\name{number.srs}
\alias{number.srs}
\title{Returns the number of spectrometers attach to the computer
It is the number of spectrometers that can be addressed}
\usage{
  number.srs(jwrapper)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}
}
\description{
  Returns the number of spectrometers attach to the
  computer It is the number of spectrometers that can be
  addressed
}

