% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{onLoad}
\alias{onLoad}
\alias{.onLoad}
\title{This private function automatically loads Omnidriver}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{name of library (automatically supplied)}

\item{pkgname}{name of the package (automatically supplied)}
}
\value{
a numeric value
}
\description{
This private function automatically loads Omnidriver
}
\details{
Package 'rOmniDriver' is a wrapper on a Java runtime available as
   a .jar file. During loading the location of the .jar file needs to be 
   found and loaded. During installation of 'rOmniDriver' a system
   environment variable is normally set to the path. However, there is
   no guarantee that it exists or that OmniDriver is installed. Consequently,
   in case of failure a diagnosis of the problem is attempted and a 
   message issued.
}
\note{
To force display of diagnosis message even when loading succeeds use 
  \code{Sys.setenv(OOI_DIAG = "TRUE")} before loading the package. This
  can be usefull to find out where the OmniDriver Java code is installed.
}
\seealso{
see also \code{\link[rJava]{jpackage}}.
}
\keyword{internal}
