% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.api.R
\name{oo_method_exists}
\alias{oo_method_exists}
\title{Check if names are available through the Java wrapper}
\usage{
oo_method_exists(jwrapper, method.name, value = TRUE)
}
\arguments{
\item{jwrapper}{jobjRef A Java Wrapper object from OmniDriver.}

\item{method.name}{character vector The name of the method(s) to search for.}

\item{value}{logical If \code{TRUE} names are returned as character strings,
otherwise numeric indexes into the matching names in \code{method.name}.}
}
\value{
A character vector, containing the methods from \code{method.names}
  that are available.
}
\description{
OmniDriver shows some variation with respect to the methods it exposes
through the wrapper. This test makes it possible to avoid Java errors
and possibly use an alternative method or issue a message.
}
\keyword{internal}
