% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-to-string.R
\name{oo_timestamp_to_string}
\alias{oo_timestamp_to_string}
\title{Convert time stamp to string}
\usage{
oo_timestamp_to_string(jwrapper, oo_timestamp)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{oo_timestamp}{HighResTimeStamp defined in OmniDriver API and retrieved
or constructed with other OmniDriver API methods. Not vectorized!}
}
\value{
a character string.
}
\description{
Convert a time stamp into a character string. Times with millisecond 
resolution and nanosecond resolutions are separate, with the nanosecond
component useful for computing time differences only.
}
\note{
Nanosecond timing and normal timing are not coincident.
}
\seealso{
Other high speed acquisition functions: 
\code{\link{get_seconds_time_delta}()},
\code{\link{get_seconds_time_delta_since}()},
\code{\link{highSpdAcq_allocate_buffer}()},
\code{\link{highSpdAcq_get_number_of_spectra_acquired}()},
\code{\link{highSpdAcq_get_spectrum}()},
\code{\link{highSpdAcq_get_time_stamp}()},
\code{\link{highSpdAcq_start_acquisition}()}
}
\concept{high speed acquisition functions}
