% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.all.spectrometers.R
\name{open_all_spectrometers}
\alias{open_all_spectrometers}
\alias{number_srs}
\title{Opens connections to all spectrometers}
\usage{
open_all_spectrometers(jwrapper)

number_srs(jwrapper)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}
}
\value{
A numeric value giving the number of USB-connected spectrometers.
}
\description{
Opens connections to communicate with all Ocean Optics spectrometers
connected through USB and returns the number of spectrometers that can be
addressed.
}
