\name{rOO-package}
\alias{rOO-package}
\alias{rOO}
\docType{package}
\title{
Package to control OO spectrometer from within R.
}
\description{
Use the Omini Driver from Ocean Optics, written in Java, from R by means of rJava.
}
\details{
\tabular{ll}{
Package: \tab rOO\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-19\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
##~~ simple examples of the most important functions ~~
}
