\name{set.boxcar.width}
\alias{set.boxcar.width}
\title{Set setting "boxcar width" (the number of pixels averaged)}
\usage{
  set.boxcar.width(jwrapper, half.width.px = 1L,
    sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{half.width.px}{the number of pixels in one half of
  the bocar window (an integer number)}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Set setting "boxcar width" (the number of pixels
  averaged)
}

