\name{set.correct.for.detector.nonlinearity}
\alias{set.correct.for.detector.nonlinearity}
\title{Set setting "correct for detector nonlinearity" (the number of pixels averaged)
for addressed spectrometer}
\usage{
  set.correct.for.detector.nonlinearity(jwrapper,
    correct = 0L, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{correct}{0: no, 1: yes}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Set setting "correct for detector nonlinearity" (the
  number of pixels averaged) for addressed spectrometer
}

