\name{set.correct.for.electrical.dark}
\alias{set.correct.for.electrical.dark}
\title{Set setting "correct for electrical dark signal"
for addressed spectrometer}
\usage{
  set.correct.for.electrical.dark(jwrapper, enable = 0L,
    sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{enable}{0: no, 1: yes}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Set setting "correct for electrical dark signal" for
  addressed spectrometer
}

