\name{set.integration.time}
\alias{set.integration.time}
\title{Set "integration time" for addressed spectrometer}
\usage{
  set.integration.time(jwrapper, time.usec = 100L,
    sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{time.usec}{integartion time in microseconds (an
  integer)}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Set "integration time" for addressed spectrometer
}

