\name{set.scans.to.avg}
\alias{set.scans.to.avg}
\title{Set "number of scans to average" for addressed spectrometer}
\usage{
  set.scans.to.avg(jwrapper, n.scans = 1L, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{n.scans}{number of scans (an integer)}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Set "number of scans to average" for addressed
  spectrometer
}

