% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.boxcar.width.R
\name{set_boxcar_width}
\alias{set_boxcar_width}
\title{Set "boxcar width"}
\usage{
set_boxcar_width(jwrapper, half.width.px = 0L, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{half.width.px}{the number of pixels in one half of the bocar window (an
integer number)}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more 
than one channel.}
}
\value{
a numeric value
}
\description{
Set setting "boxcar width" (the number of adjacent pixels averaged)
}
\note{
The signature of this function is different to that of the Java
  equivalent, so that as with other functions in the package indexes for
  addressing spectrometer and channel can have default values.
}
