% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.correct.for.detector.nonlinearity.R
\name{set_correct_for_detector_nonlinearity}
\alias{set_correct_for_detector_nonlinearity}
\title{Set "correct for detector nonlinearity"}
\usage{
set_correct_for_detector_nonlinearity(
  jwrapper,
  correct = 0L,
  sr.index = 0L,
  ch.index = 0L
)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{correct}{integer or logical; 0, FALSE: apply correction; 1, TRUE: do 
not apply correction.}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more 
than one channel.}
}
\value{
a logical value that the user SHOULD test as the method may fail.
}
\description{
Set setting "correct for detector nonlinearity" for the addressed
spectrometer.
}
\note{
The signature of this function is different to that of the Java 
  equivalent, so that as with other functions in the package indexes for 
  addressing spectrometer and channel can have default values.
}
