% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.correct.for.electrical.dark.R
\name{set_correct_for_electrical_dark}
\alias{set_correct_for_electrical_dark}
\title{Set "correct for electrical dark signal"}
\usage{
set_correct_for_electrical_dark(
  jwrapper,
  enable = 0L,
  sr.index = 0L,
  ch.index = 0L
)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{enable}{integer or logical; 0, FALSE: disable; 1, TRUE: enable.}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more 
than one channel.}
}
\value{
a logical value
}
\description{
Set setting "correct for electrical dark signal"
for the addressed spectrometer channel
}
\note{
The signature of this function is different to that of the Java
  equivalent, so that as with other functions in the package indexes for
  addressing spectrometer and channel can have default values.
}
