% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.integration.time.R
\name{set_integration_time}
\alias{set_integration_time}
\title{Set "integration time"}
\usage{
set_integration_time(jwrapper, time.usec = 100L, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{time.usec}{integartion time in microseconds (an integer)}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more 
than one channel.}
}
\value{
a numeric value
}
\description{
Set "integration time" for the addressed spectrometer channel
}
\note{
The signature of this function is different to that of the Java
  equivalent, so that as with other functions in the package indexes for
  addressing spectrometer and channel can have default values.
}
