% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.scans.to.avg.R
\name{set_scans_to_average}
\alias{set_scans_to_average}
\alias{set_scans_to_avg}
\title{Set "number of scans to average"}
\usage{
set_scans_to_average(jwrapper, n.scans = 1L, sr.index = 0L, ch.index = 0L)

set_scans_to_avg(jwrapper, n.scans = 1L, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{n.scans}{number of scans (an integer)}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more 
than one channel.}
}
\value{
a numeric value
}
\description{
Set "number of scans to average"
}
\note{
The signature of this function is different to that of the Java
  equivalent, so that as with other functions in the package indexes for
  addressing spectrometer and channel can have default values.
}
