% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-spectrum-type.R
\name{set_spectrum_type}
\alias{set_spectrum_type}
\title{Set "spectrum type"}
\usage{
set_spectrum_type(spct.type.wrapper, spct.type = "normal")
}
\arguments{
\item{spct.type.wrapper}{feature wrapper as returned by function
\code{rOmniDriver::get_feature_controller_internal_trigger}}

\item{spct.type}{character, "normal" or "raw" or interger, 0 or 1.}
}
\value{
logical TRUE = success, FALSE = failure.
}
\description{
Set the spectrum type to one of normal  or raw.
}
