% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.timeout.R
\name{set_timeout}
\alias{set_timeout}
\title{Set (trigger) "timeout"}
\usage{
set_timeout(jwrapper, time.millisec = 1000L, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{time.millisec}{time to wait in milliseconds (an integer)}

\item{sr.index}{an index to address the spectrometer}
}
\value{
a numeric value
}
\description{
Setting a "timeout" for the addressed spectrometer is useful when used
together with triggers to avoid endless waiting in case the trigger input
does not take place.
}
\note{
This method is applied to all channels.
}
