\name{spectrum.flush}
\alias{spectrum.flush}
\title{Flush the last spectrum acquired or currently being acquired}
\usage{
  spectrum.flush(jwrapper, sr.index = 0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
  Flush the last spectrum acquired or currently being
  acquired
}

