% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.flush.R
\name{spectrum_flush}
\alias{spectrum_flush}
\title{Flush the most recent spectrum}
\usage{
spectrum_flush(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
a numeric value
}
\description{
Flush the most recent spectrum acquired or currently being acquired
}
