% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop.averaging.R
\name{stop_averaging}
\alias{stop_averaging}
\title{Stop averaging of scans}
\usage{
stop_averaging(jwrapper, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer 
for the time being not exported}

\item{ch.index}{an index to address the channel in a spectrometer with more 
than one channel.}
}
\value{
a numeric value
}
\description{
Stop averaging for addressed spectrometer probably not very useful in
"vanilla" R as this aborts the current averaging of spectra, but needs a
multithreaded application to be able to call it before get.spectrum returns.
}
\note{
The signature of this function is different to that of the Java
  equivalent, so that as with other functions in the package indexes for
  addressing spectrometer and channel can have default values.
}
