% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.bench.R
\name{get_bench}
\alias{get_bench}
\title{Get optical properties of optical bench}
\usage{
get_bench(jwrapper, sr.index = 0L, ch.index = NULL)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a value
}
\description{
A function to query the characteristics of the optical bench
from a spectrometer.
}
\keyword{manip}
\keyword{misc}

