% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.boxcar.width.R
\name{get_boxcar_width}
\alias{get_boxcar_width}
\title{Get current setting of boxcar width}
\usage{
get_boxcar_width(jwrapper, sr.index = 0L, ch.index = NULL)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a numeric value
}
\description{
Function to query the boxcar width being used
(the number of pixels averaged).
}
\keyword{manip}
\keyword{misc}

