\name{get_correct_for_stray_light}
\alias{get_correct_for_stray_light}
\title{Get current setting of stray light correction}
\usage{
get_correct_for_stray_light(jwrapper, sr.index=0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\description{
Get current setting of stray light correction
}

