\name{get_detector}
\alias{get_detector}
\title{Get description of detector}
\usage{
get_detector(jwrapper, sr.index=0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\value{
a numeric value
}
\description{
Get description of detector
}
\keyword{manip}
\keyword{misc}

