% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.feature.controller.board.temperature.R
\name{get_feature_controller_board_temperature}
\alias{get_feature_controller_board_temperature}
\title{Get feature "controller board temperature"}
\usage{
get_feature_controller_board_temperature(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
an object which provides the desired interface, or null if this
  feature is not available for this spectrometer.
}
\description{
A function to retrieve an interface that allows to query controller board
temperature.
}
\note{
Before calling this function you first need to check that the feature
  is supported by the spectrometer in use by calling 
  \code{\link{is_feature_supported_board_temperature}}.
}
\keyword{manip}
\keyword{misc}

