\name{get_feature_controller_board_temperature}
\alias{get_feature_controller_board_temperature}
\title{Get feature "controller board temperature"
Before calling this function you first need to check that the feature is supported
by the spectrometer in use.}
\usage{
get_feature_controller_board_temperature(jwrapper, sr.index=0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\value{
a numeric value
}
\description{
Get feature "controller board temperature" Before calling
this function you first need to check that the feature is
supported by the spectrometer in use.
}
\keyword{manip}
\keyword{misc}

