% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.feature.pixel.binning.R
\name{get_feature_pixel_binning}
\alias{get_feature_pixel_binning}
\title{Get feature "pixel binning"}
\usage{
get_feature_pixel_binning(jwrapper, sr.index = 0L)
}
\arguments{
\item{jwrapper}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer}
}
\value{
an object which provides the desired interface, or null if this
  feature is not available for this spectrometer.
}
\description{
Obtain an inferface to the "pixel binning" function of a spectrometer if
available.
}
\note{
Before calling this function you first need to check that the feature
  is supported by the spectrometer in use by calling
  \code{\link{is_feature_supported_pixel_binning}}.
}
\keyword{manip}
\keyword{misc}

