\name{get_spectrum}
\alias{get_spectrum}
\title{Get a spectrum from the spectrometer}
\usage{
get_spectrum(jwrapper, sr.index=0L)
}
\arguments{
  \item{jwrapper}{an open Wrapper object from Omnidriver}

  \item{sr.index}{an index to address the spectrometer}
}
\value{
a numeric value
}
\description{
Get a spectrum from the spectrometer
}
\keyword{manip}
\keyword{misc}

