\docType{package}
\name{rOmniDriver-package}
\alias{rOmniDriver}
\alias{rOmniDriver-package}
\title{Use Ocean Optics' Omnidriver from R}
\description{
A package to use OmniDriver (a library from Ocean Optics)
}
\details{
OmniDriver is a propietary library provided by Ocean Optics
to allow the control of the spectrometers they sell through
third party software.

Although this package is under GPL, Omnidriver is not, it
is copyrighted commercial software from Ocean Optics
\url{http://www.oceanoptics.com}

Ocean optics has authorized the redistribution of the parts
of OmniDriver needed for this package to work, free of
charge.

Ocean Optics will not provide support for the use of this
package, unless the user buys a licence for OmniDriver SDP
(spectroscopy development plattform). All support requests
about this package should be addressed to the maintainer of
the package or public R user forums.
}
\author{
Pedro J. Aphalo
}
\references{
\url{http://www.oceanoptics.com/Products/omnidriver.asp}
}
\keyword{package}

